requirements.txt
SQLAlchemy~=2.0.36
flask~=3.1.0
flask-login~=0.6.3
flask-sqlalchemy~=3.1.1
werkzeug~=3.1.3
python-dotenv~=1.0.1
mysql-connector-python~=9.1.0
